backend/requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
fastapi==0.124.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.0
httptools==0.7.1
idna==3.11
lxml==6.0.2
mako==1.3.10
markdown==3.10
markupsafe==3.0.3
numpy==2.3.5
pdfminer-six==20251107
pdfplumber==0.11.8
pgvector==0.4.2
pillow==12.0.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.3.2
psycopg-binary==3.3.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pyparsing==3.2.5
pypdf2==3.0.1
pypdfium2==5.1.0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pyyaml==6.0.3
requests==2.32.5
rsa==4.9.1
sqlalchemy==2.0.44
starlette==0.50.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
